backend/requirements.txt
# EllinCRM - Backend Dependencies
# Python 3.13+

# Web Framework
fastapi>=0.128.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Data Validation
pydantic>=2.12.0
pydantic-settings>=2.8.0
email-validator>=2.2.0

# Database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.30.0
alembic>=1.14.0

# Data Processing
beautifulsoup4>=4.12.0
lxml>=5.3.0
openpyxl>=3.1.0
python-dateutil>=2.9.0
pdfplumber>=0.11.0

# AI/ML (Semantic Search & Embeddings)
sentence-transformers>=3.4.0
pgvector>=0.3.0
numpy>=1.26.0

# LLM Router
litellm>=1.81.0

# Logging
structlog>=24.4.0

# HTTP Client
httpx>=0.28.0

# Google Sheets Integration
google-api-python-client>=2.154.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Testing
pytest>=8.3.0
pytest-cov>=6.0.0
pytest-asyncio>=0.24.0

# Code Quality
ruff>=0.8.0
mypy>=1.13.0

# Types
types-python-dateutil>=2.9.0
